refactor : community entity
